% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angelopoulos_original.R
\name{angelopoulos_original}
\alias{angelopoulos_original}
\title{PPI Linear Regression using Angelopoulos et al. (2023) Analytic Form}
\usage{
angelopoulos_original(rec_form, inf_form, dat)
}
\arguments{
\item{rec_form}{A formula defining the rectifier model. This should be of
the form Y - Yhat ~ X, where Y is the name of the column corresponding to
the observed outcome in the labeled data, Yhat is the name of the column
corresponding to the predicted outcome in the labeled data, and X generally
corresponds to the features of interest (e.g., X1 + X2).}

\item{inf_form}{A formula defining the inference model. This should be of
the form Yhat ~ X, where Yhat is the name of the column corresponding to the
predicted outcome in the unlabeled data, and X generally corresponds to the
features of interest (e.g., X1 + X2).}

\item{dat}{data in the form of the simdat function output}
}
\value{
a list with linear regression parameter estimate and standard error.
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{

rec_form <- Y - Yhat ~ X1

inf_form <- Yhat ~ X1

dat <- simdat()

angelopoulos_original(rec_form, inf_form, dat = dat)

}
