% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{ppi}
\alias{ppi}
\title{Valid and Efficient Post-Prediction Inference (PPI) using State-of-the-Art
Methods}
\usage{
ppi(
  formula,
  estimand,
  method,
  data,
  set,
  seed = NULL,
  alpha = 0.05,
  alternative = "two-sided",
  ...
)
}
\arguments{
\item{formula}{description ... of the form Y - f ~ X1 + X2 + ...}

\item{estimand}{description}

\item{method}{description}

\item{data}{description}

\item{set}{... column name for lab versus unlab...}

\item{seed}{description}

\item{alpha}{description}

\item{alternative}{description}

\item{...}{Further arguments passed to or from specific methods. See details.}
}
\value{
description
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{
# example code


}
