% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusplus.R
\name{ppi_ols_ci}
\alias{ppi_ols_ci}
\title{PPI++ Linear Regression using Angelopoulos et al. (2023) Analytic Form}
\usage{
ppi_ols_ci(
  rec_form,
  inf_form,
  dat,
  alpha = 0.05,
  alternative = "two-sided",
  lhat = NULL,
  coord = NULL,
  w = NULL,
  w_unlabeled = NULL
)
}
\arguments{
\item{rec_form}{A formula defining the rectifier model. This should be of
the form Y - Yhat ~ X, where Y is the name of the column corresponding to
the observed outcome in the labeled data, Yhat is the name of the column
corresponding to the predicted outcome in the labeled data, and X generally
corresponds to the features of interest (e.g., X1 + X2).}

\item{inf_form}{A formula defining the inference model. This should be of
the form Yhat ~ X, where Yhat is the name of the column corresponding to the
predicted outcome in the unlabeled data, and X generally corresponds to the
features of interest (e.g., X1 + X2).}

\item{dat}{data in the form of the simdat function output}

\item{alpha}{scalar type I error rate for hypothesis testing - values in (0, 1); defaults to 0.05}

\item{alternative}{alternative hypothesis for hypothesis testing - options include "one-sided" or "two-sided"; defaults to two-sided}

\item{lhat}{scalar power tuning parameter - values in (0, 1); defaults to NULL and is estimated}

\item{coord}{vector coordinates for gradient descent; defaults to NULL and is estimated}

\item{w}{vector weights for weighted glm in labeled dataset; defaults to NULL and is estimated}

\item{w_unlabeled}{vector weights for weighted glm in unlabeled dataset; defaults to NULL and is estimated}
}
\value{
A list of outputs: estimate of inference model parameters and corresponding standard errors
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{

rec_form <- Y - Yhat ~ X1

inf_form <- Yhat ~ X1

dat <- simdat()

ppi_ols_ci(rec_form, inf_form, dat = dat)

}
