% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predpi_ols.R
\name{predpi_ols}
\alias{predpi_ols}
\title{PPI Linear Regression using Angelopoulos et al. (2023) Analytic Form}
\usage{
predpi_ols(X_l, Y_l, f_l, X_u, f_u, n, p, N, ...)
}
\arguments{
\item{rec_form}{A formula defining the rectifier model. This should be of     # Redo these to conform to new conventions
the form Y - Yhat ~ X, where Y is the name of the column corresponding to
the observed outcome in the labeled data, Yhat is the name of the column
corresponding to the predicted outcome in the labeled data, and X generally
corresponds to the features of interest (e.g., X1 + X2).}

\item{inf_form}{A formula defining the inference model. This should be of
the form Yhat ~ X, where Yhat is the name of the column corresponding to the
predicted outcome in the unlabeled data, and X generally corresponds to the
features of interest (e.g., X1 + X2).}

\item{dat}{data in the form of the simdat function output}
}
\value{
A list of outputs: estimate of inference model parameters and corresponding standard errors
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{

rec_form <- Y - Yhat ~ X1

inf_form <- Yhat ~ X1

dat <- simdat()

angelopoulos_original(rec_form, inf_form, dat = dat)         ## Update with wrapper function or keep method-specific examples?

}
