% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_analytic.R
\name{wang_analytic}
\alias{wang_analytic}
\title{PPI Linear Regression using Wang et al. (2020) Analytic Correction}
\usage{
wang_analytic(rel_form, inf_form, dat)
}
\arguments{
\item{rel_form}{A formula defining the relationship model. This should be
of the form Y ~ Yhat, where Y is the name of the column corresponding to
the observed outcome in the labeled data and Yhat is the name of the column
corresponding to the predicted outcome in the labeled data.}

\item{inf_form}{A formula defining the inference model. This should be of
the form Yhat ~ X, where Yhat is the name of the column corresponding to the
predicted outcome in the unlabeled data, and X generally corresponds to the
features of interest (e.g., X1 + X2).}

\item{dat}{data in the form of the simdat function}
}
\value{
A list of outputs: estimate of inference model parameters and corresponding standard error.
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{

rel_form <- Y ~ Yhat

inf_form <- Yhat ~ X1

dat <- simdat()

wang_analytic(rel_form, inf_form, dat = dat)

}
